[Refactor] 에러 핸들링 컴포넌트 분리 #5
