+++ src/pages/map/components/index.ts
@@ -1,1 +1,2 @@
 export { default as SearchBar } from "./SearchBar";
+export { default as Error } from "./Error";
